refactor(script): ページネーション構成スクリプトを共通化
